msm: ipa3: Add smmu mapping to proper context bank
For wdi1 protocol, smmu mapping of buffer of MCC pipes
happens in WLAN context bank. With the introduction of
new API in latest kernel version for replenishing
rx_cache, buffer mapping of MCC pipes was happening
in AP context bank which is leading to crash. Now
changing smmu mapping of buffer in WLAN context bank.

Change-Id: I208c1e1eae5b103ca911630c37e42965ad1397f7
Signed-off-by: Piyush Dhyani<quic_pdhyani@quicinc.com>