msm: ipa3: new header insertion mechanism for IPAv6.0
Introducing new Design to meet the IPA6.0 requirement
to support 128 clients with 256 headers in SRAM for peak performance.

New SRAM Section “apps_headers_ext” will be used to store the headers
when the apps_headers section is full. HPC in SRAM will be used to
access the new section through routing rule.

Change-Id: Iee4ee38478f1d48fa9d224cfa62a32f920621886