msm: ipa3: Adding prefetch buff size for kalama target
Due to missing the prefetch buff size route and filter table will
overflow if IPA HW tries to prefetch 128Bytes of additional memory when
dealing with filter/routing tables. This can result in overflowing
of memory if we map only the required size. Make changes to
add the prefetch buff size for kalama target.

Change-Id: I045e14b25e397dc800a775f0d5cc02a889fc7c21
Signed-off-by: Ashok Vuyyuru <quic_avuyyuru@quicinc.com>