msm: ipa3: Adding prefetch buff size for sdxpinn target
Due to missing the prefetch buff size route and filter table will
overflow if IPA HW tries to prefetch 128Bytes of additional memory when
dealing with filter/routing tables. This can result in overflowing
of memory if we map only the required size. Make changes to
add the prefetch buff size for sdxpinn target.

Change-Id: I7544e4aced009592c50622410a6f74a88b05f92a
Signed-off-by: Ashok Vuyyuru <quic_avuyyuru@quicinc.com>