dataipa: Fix IPv6CT table switch failure in few iterations
Issue: When the CT table is moved between SRAM or DDR, we observe failure in few iterations when the table had an entry and was
previously deleted. But we still try to look up for handles for
the deleted entry.

Fix: There could be cases when the entry needs to be deleted, but
we don't change enabled bit to 0, but change protocol to 0xFF,
so we should compare against 0xFF instead of 0xFF00.

Change-Id: I714c9f4084708af3f0b55e415d6ab64394ef9782