ipa: Protect XFRM callbacks from non-clocked accesses
We must guaranty that XFRM callbacks, which access the HW
will do so with IPA clocks on.
1. Adding regular voting for the relaxed callbacks.
2. Since the ipa_ipsec_xdo_state_update_curlft may be called
from the timer interrupt, we must not use the regular API,
which invokes mutex. Instead, like in the IPA interrup handler,
we will update the SA life bytes only if the IPA is clocked.

Change-Id: Icbe5deb64dc279d3923ab40a3b42d6dbe6d16700
Signed-off-by: Ilia Lin <quic_ilial@quicinc.com>