data-ipanat: Increase number of entries in expansion table
Issue: On testing with max supported connections which is 4096
currently we observe a lot of hash collisions leading to packets
taking SW path as expansion table is completely filled.

Fix: Added chnage to increase the number of expansion table
entries from 800 to (max_entries configured in IPACM_cfg.xml)/2 to support 4096 parallel connections.

Change-Id: I9ac21e7a19862b3633cd19505e5a9be02e7eaffe